Fix 128-bit integer support and memory pool alignment
- Fix CXY_AST_NODE_BODY_SIZE macro to use offsetof() for correct padding calculation
- Replace sputu128/sputi128 with chunked sprintf approach in builtins.cxy
- Add INT128_HIGH/INT128_LOW/INT128 macros for 128-bit manipulation
- Resolves memory corruption issue when clearAstBody was called with 128-bit integers
- All tests passing with robust 128-bit integer formatting

+++ src/cxy/core/sb.h
@@ -64,8 +64,8 @@ static attr(always_inline) void stringBuilderAppendChar(StringBuilder *sb,
     stringBuilderAppendCstr0(sb, s.str, s.str[5]);
 }
 
-void stringBuilderAppendInt(StringBuilder *sb, i64 num);
-void stringBuilderAppendUInt(StringBuilder *sb, u64 num);
+void stringBuilderAppendInt(StringBuilder *sb, i128 num);
+void stringBuilderAppendUInt(StringBuilder *sb, u128 num);
 void stringBuilderAppendFloat(StringBuilder *sb, f64 num);
 void stringBuilderAppendBool(StringBuilder *sb, bool v);
 char *stringBuilderRelease(StringBuilder *sb);
